(delayedScatter) BUGFIX & feature. fitTime now IS myFitTime (instead of hard-coded wrong value). Also added feature: determination of dY_time field. Small minor changes.

+++ Martijn_custom/CRPcAMP_preliminary_delayedScatterAndPlot.m
@@ -7,6 +7,14 @@
 % NW's data handling; 
 % MW TODO: Check PN data.
 %
+% About input params:
+% - associatedFieldNames{1} MUST BE timefield field name
+% - associatedFieldNames{2} X, usually fluor concentration or rate field name
+% - associatedFieldNames{3} Y, usually growth rate field name
+%
+% - makeDtime   if set to 0 nothing happens, if set to 1, extra 
+%               MWDJK_dX_time filed is recalculated.
+%
 % Example of how to call script:
 %{
 myID = ' icd-lac';
@@ -42,52 +50,97 @@ if ~exist('myOutputFolder')
     myOutputFolder = 'C:\Users\wehrens\Desktop\testdelayedscatter\output\';
 end
 
-if ~exist('myDataFile') | ~exist('associatedFieldNames') | ~exist('myTitle') | ~exist('p') | ~exist('badSchnitzes')
+if ~exist('associatedFieldNames') | ~exist('myTitle') | ~exist('p') | ~exist('badSchnitzes')
     error('input not supplied.')
 end
     
 
 % Loading
-load(myDataFile);
+%load(myDataFile); % Can also be done by user
 
 
 %%
 
-
-% Find Schnitzes with slow/negative growth rate -> rm them?!
-slowschnitzes=NW_detectSlowSchnitzes(p,schnitzcells,'muP15_fitNew','muThreshold',0.05);
-
-% Preparation to load
-% Adapted from NW excel sheet
+name_rm = 'rm'; name_all = 'all';
 fitTime = myFitTime;
-s_all = DJK_selSchitzesToPlot(schnitzcells, 'P', @(x) 1); name_all = 'all';
-s_all_fitTime = DJK_selSchitzesToPlot(s_all, 'time', @(x) x(1) > fitTime(1) & x(1) < fitTime(2)); name_all_fitTime = ['all_' num2str(fitTime(1)) '_' num2str(fitTime(2))];
-s_all_fitTime_cycle = DJK_selSchitzesToPlot(s_all_fitTime, 'completeCycle', @(x) x ~= 0); name_all_fitTime_cycle = [name_all_fitTime '_cycle'];
 
-s_rm = DJK_selSchitzesToPlot(s_all, 'P', @(x) 1); name_rm = 'rm';
-for i=badSchnitzes, s_rm(i).useForPlot=0; end;
-s_rm_fitTime = DJK_selSchitzesToPlot(s_rm, 'time', @(x) x(1) > fitTime(1) & x(1) < fitTime(2)); name_rm_fitTime = ['rm_' num2str(fitTime(1)) '_' num2str(fitTime(2))];
-s_rm_fitTime_cycle = DJK_selSchitzesToPlot(s_rm_fitTime, 'completeCycle', @(x) x ~= 0); name_rm_fitTime_cycle = [name_rm_fitTime '_cycle'];
+if ~alreadyRemovedInMatFile
+    % Find Schnitzes with slow/negative growth rate -> rm them?!
+    slowschnitzes=NW_detectSlowSchnitzes(p,schnitzcells,associatedFieldNames{3},'muThreshold',0.05);
 
+    % Preparation to load
+    % Adapted from NW excel sheet    
+    s_all = DJK_selSchitzesToPlot(schnitzcells, 'P', @(x) 1); 
+    s_all_fitTime = DJK_selSchitzesToPlot(s_all, 'time', @(x) x(1) > fitTime(1) & x(1) < fitTime(2)); name_all_fitTime = ['all_' num2str(fitTime(1)) '_' num2str(fitTime(2))];
+    s_all_fitTime_cycle = DJK_selSchitzesToPlot(s_all_fitTime, 'completeCycle', @(x) x ~= 0); name_all_fitTime_cycle = [name_all_fitTime '_cycle'];
 
-%% Calculating branches
-% ===
-s_rm = MW_calculateframe_nrs(s_rm); % backwards compatibility fix 
+    s_rm = DJK_selSchitzesToPlot(s_all, 'P', @(x) 1); 
+    for i=badSchnitzes, s_rm(i).useForPlot=0; end;
+    s_rm_fitTime = DJK_selSchitzesToPlot(s_rm, 'time', @(x) x(1) > fitTime(1) & x(1) < fitTime(2)); name_rm_fitTime = ['rm_' num2str(fitTime(1)) '_' num2str(fitTime(2))];
+    s_rm_fitTime_cycle = DJK_selSchitzesToPlot(s_rm_fitTime, 'completeCycle', @(x) x ~= 0); name_rm_fitTime_cycle = [name_rm_fitTime '_cycle'];
+end
 
-fitTime = [100 630]; fitTime = fitTime + [2 -2];
-
+%% 
 % Collect indices of the GFP measurement per schnitz.
 % This is not necessary because Noreen's data already contains the right
 % fields -and the fluor code already does this 
 % (DJK_addToSchnitzes_fluor_anycolor) - but might be convenient for future 
 % generation of "fieldX_at_G".
-indicesForG = {};
-for i = 1:numel(s_rm)    
-    indicesForG{i} = find(~isnan(s_rm(i).G_mean_all));
+indicesForFluor = {};
+FluorFieldAllName = [upper(p.fluor1) '_mean_all'];
+FluorRateBaseFieldAllName = [upper(p.fluor1) '_time'];
+numelS_rm = numel(s_rm);
+for i = 1:numelS_rm
+    % current schnitz
+    s = s_rm(i);
+    
+    % Re-calculate fluor times
+    indicesForFluor{i} = find(~isnan(s_rm(i).(FluorFieldAllName))); % could access field name as string
+    % This is only valid for concentration values:
+    theTimes = s_rm(i).('time');
+    s_rm(i).MWTimeField = theTimes(indicesForFluor{i});
+    
+    % Calculate dX_time field if it doesn't exist, based on the assumption
+    % it was made by DJK_addTocSchnitzes_fluorRate_phase, which takes into
+    % acount the points t+1 and t.
+    if makeDtime        
+            
+            % include the daughter field that was also used for fluor calc
+            if s.D ~= 0
+                % take only one daughter for time reference
+                sD = s_rm(s.D);
+                % extend
+                extendedTime = [s.(FluorRateBaseFieldAllName), sD.(FluorRateBaseFieldAllName)(1)];
+            else
+                % end of lineage, no extension
+                extendedTime = [s.(FluorRateBaseFieldAllName)]; 
+            end
+            
+            % calculate times
+            % sum n and n+1, average
+            DTimes = [extendedTime(1:end-1)+extendedTime(2:end)]./2;
+                    
+            % add to schnitz
+            fieldName = ['MWDJK_' 'd' FluorRateBaseFieldAllName];
+            s_rm(i).(fieldName) = DTimes;
+
+    end
 end
 
+if makeDtime
+    warning('Determination of dX_time values only valid for DJK_addToSchnitzes_fluor_anycolor rates.');
+    pause(1);
+end
+
+%% Calculating branches
+% ===
+s_rm = MW_calculateframe_nrs(s_rm); % backwards compatibility fix 
+
+fitTime = fitTime + [2 -2];
+
+
 branchData = DJK_getBranches(p,s_rm,'dataFields',{associatedFieldNames{1}, associatedFieldNames{2}, associatedFieldNames{3} }, 'fitTime', fitTime); 
-name_rm_branch = [name_rm '_' num2str(fitTime(1)) '_' num2str(fitTime(2)) '_Conc_oldRates'];
+ name_rm_branch = [name_rm '_' num2str(fitTime(1)) '_' num2str(fitTime(2)) '_Conc_oldRates'];
 
 
 %%
@@ -103,6 +156,7 @@ for branch_nr = 1:numelBranches
     set(l, 'LineWidth', (numelBranches-branch_nr+1)/numelBranches*10);
 end
 
+xlabel(associatedFieldNames{1},'Interpreter', 'None'), ylabel(associatedFieldNames{3},'Interpreter', 'None')
 
 %%
 
@@ -120,7 +174,7 @@ p.extraNorm=0;
 % THIS MIGHT FAIL BECAUSE TIME IS NOT SET CORRECTLY (SHOULD BE time_at_..)
 % To calculate cross correlations, additional normalization is usually
 % performed, namely to filter out colony average behavior.
-[CorrData,composite_corr] = DJK_plot_crosscorrelation_standard_error_store(p, branch_groups, ['noise_' associatedFieldNames{1,2}],['noise_' associatedFieldNames{1,3}] ,'selectionName',name_rm_branch,'timeField',associatedFieldNames{1},'onScreen',0); 
+[CorrData,composite_corr] = DJK_plot_crosscorrelation_standard_error_store(p, branch_groups, ['noise_' associatedFieldNames{1,2}],['noise_' associatedFieldNames{1,3}] ,'selectionName',name_rm_branch,'timeField',associatedFieldNames{1},'onScreen',1); 
 
 % Do we want to filter out colony average behavior for the "delayed
 % scatter" plots also? Maybe do this with noise fields?
@@ -134,14 +188,16 @@ if isfield(p,'tauIndices'), p=rmfield(p,'tauIndices'); end
 %% Compare two cross-corrs (DJK & MW)
 
 figure(1),clf,hold on
-plot(CorrData(:,1),CorrData(:,2),'x-k','LineWidth',2)
-plot(CorrData(:,1),correlationsPerTau,'o-r','LineWidth',2)
+errorbar(CorrData(:,1),CorrData(:,2),CorrData(:,3),'x-','Color', [.5,.5,.5], 'LineWidth',2)
+l1=plot(CorrData(:,1),CorrData(:,2),'x-k','LineWidth',2)
+
+l2=plot(CorrData(:,1),correlationsPerTau,'o-r','LineWidth',2)
 myxlimvalues=[min(CorrData(:,1)), max(CorrData(:,1))];
 xlim(myxlimvalues);
-ylim([-0.2,0.2]);
+ylim([-0.4,0.4]);
 plot(myxlimvalues,[0,0],'k-');
 
-legend({'DJK','MW'})
+legend([l1,l2],{'DJK','MW'})
 
 title(['DJK vs. MW R -- ' myID '_' p. movieDate  '_' p.movieName], 'Interpreter', 'none');
 xlabel('\tau (hrs)');
@@ -242,3 +298,39 @@ if PLOT3DSCATTER
     %xlim([0, max([myData(:).selected_growth_rates])])
 end
     
+
+%% Some random checks
+
+% distribution of times 
+% equivalent of 'hist' calculated manually because times have interval,
+% which complicates use of hist function.
+fieldcount=0; R = struct;
+% loop over timefields of interest
+for myfield = {'Y_time','MWDJK_dY_time', 'time'}
+
+    % administration
+    myfield = char(myfield);
+    fieldcount = fieldcount+1;
+    
+    % get all possible values 
+    R(fieldcount).YValues = unique([s_rm.(myfield)]);
+    % count how many of each values are in the schnitz struct
+    R(fieldcount).countYValues = [];
+    for value=R(fieldcount).YValues 
+        R(fieldcount).countYValues(end+1) = sum([s_rm.(myfield)]==value);
+    end
+    
+end
+% plot first two fields of interest
+figure, clf; hold on;
+if numel(R)>2, plot(R(3).YValues, R(3).countYValues, 'xk'); end
+plot(R(1).YValues, R(1).countYValues, 'or','LineWidth',2);
+plot(R(2).YValues, R(2).countYValues, 'ob','LineWidth',2);
+
+
+
+
+
+
+
+
